requirements.txt
aiohttp==3.5.4
aiohttp-jwt==0.2.0
aiohttp-validate==1.0.1
aiomysql==0.0.20
aioredis==1.2.0
aiozabbix==1.0.0
amqp==2.4.0
asn1crypto==0.24.0
async-timeout==3.0.1
asyncssh==1.15.0
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
celery-once==2.0.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
cx-Oracle==7.0.0
hiredis==0.3.0
idna==2.8
iso8601==0.1.12
jsonschema==2.6.0
kombu==4.2.1
lxml==4.2.5
multidict==4.5.2
numpy==1.16.0
pandas==0.23.4
pipenv==2018.11.26
pkg-resources==0.0.0
py-zabbix==1.1.5
pyasn1==0.4.5
pyasn1-modules==0.2.2
pycparser==2.19
PyJWT==1.7.1
PyMySQL==0.9.2
python-dateutil==2.7.5
python-dotenv==0.9.1
python-engineio==3.2.1
python-ldap==3.1.0
python-socketio==2.1.0
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.16
urllib3==1.24.1
vine==1.1.4
virtualenv==16.2.0
virtualenv-clone==0.4.0
yarl==1.3.0
openpyxl==3.0.3